python_ml/requirements.txt
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.0
joblib==1.4.2
python-docx==1.1.2
python-pptx==1.0.2
PyMuPDF==1.24.0
python-dotenv==1.0.0

